pyproject.toml
[project]
name = "drafts-i"
version = "1.0.0"
description = "RAG personal assistant for my drafts"
readme = "README.md"
authors = [
    { name = "Hany Mamdouh", email = "hany@atentec.com" }
]
requires-python = ">=3.11"
dependencies = [
    "llama-index-core>=0.14.9",
    "llama-index-embeddings-huggingface>=0.6.1",
    "llama-index-readers-file>=0.5.5",
    "markdown>=3.10",
    "numpy>=2.3.5",
    "sentence-transformers>=5.1.2",
    "sqlite-vec>=0.1.6",
]

[project.scripts]
cli = "drafts_i.cli:main"
ingest = "drafts_i.ingest:main"
query = "drafts_i.query:main"

[build-system]
requires = ["uv_build>=0.9.13,<0.10.0"]
build-backend = "uv_build"

[tool.pyright]
executionEnvironments = [{ root = "src" }]
typeCheckingMode = "standard"
venv = ".venv"
venvPath = "."
